Learning how to webscrape(JC Cut-off Points scraper)
Was thinking of automating searching for a website every time i want to look for the JC cut-off points
The program does not have much features. Will be improving this program later on with more features like categorization by year and school. Will also add a save function to save data to an excel spreadsheet